pyproject.toml
[tool.poetry]
name = "run-scripts"
version = "0.1.1"
description = "A CLI for running npm-like scripts"
authors = ["Keksi <contact@keksi.me>"]
repository = "https://github.com/keksiqc/run-cli"
readme = "README.md"

[tool.poetry.scripts]
run = "run_scripts.cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
